Create Linked List - Insert Node at first
Insertion in Linked List
Let's learn about inserting new elements at the beginning of a linked list.

Now, suppose we have a linked list 1 -> 2 -> 3 and we want to insert 4 at the beginning. We can then follow these steps:

Create a new node with value 4. Let's call it newNode
Add the head of the existing linked list as the next of newNode.
Then set the head variable to the newNode, as the newNode is our new head.